Update TERM specially for FreeBSD
FreeBSD 12.x and earlier doesn't have tmux entries in its termcap. They
will exist in 13.0. Because tmux recommends using its termcap/terminfo
entries as a TERM, do that where possible. The caveat of putting a fix
in fish's config is there will be warnings when connecting over SSH
within tmux but it will be fixed immediately.